requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
playwright==1.44.0
python-dotenv==1.0.1
apscheduler==3.10.4
python-telegram-bot==21.3
# sqlite3 is a Python stdlib built-in — do NOT pip install it
fake-useragent==1.5.1
httpx==0.27.0
tenacity==8.3.0
rich==13.7.1
tabulate==0.9.0
